pyproject.toml
[tool.poetry]
name = "ai-software-factory"
version = "0.1.0"
description = "Local Agentic MVP Builder with Self-Improvement"
authors = ["Sparsh Prajapati"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
streamlit = "^1.28.0"
google-generativeai = "^0.8.5"
groq = "^0.4.0"
python-dotenv = "^1.0.0"
tavily-python = "^0.3.0"
langsmith = "^0.4.0"
pydantic = "^2.5.0"
tenacity = "^8.2.0"
pyyaml = "^6.0"
httpx = "^0.27.0"

# Optional providers (install if needed)
openai = { version = "^1.0", optional = true }
anthropic = { version = "^0.39.0", optional = true }

[tool.poetry.extras]
openai = ["openai"]
anthropic = ["anthropic"]
all-providers = ["openai", "anthropic"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
